// project-management-frontend/src/@api/models/register-company-request-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project Management Tool API
 * API documentation for Project Management Tool
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AdminDto } from './admin-dto';
import {
    AdminDtoFromJSON,
    AdminDtoFromJSONTyped,
    AdminDtoToJSON,
    AdminDtoToJSONTyped,
} from './admin-dto';

/**
 * 
 * @export
 * @interface RegisterCompanyRequestDto
 */
export interface RegisterCompanyRequestDto {
    /**
     * 
     * @type {string}
     * @memberof RegisterCompanyRequestDto
     */
    companyName?: string;
    /**
     * 
     * @type {string}
     * @memberof RegisterCompanyRequestDto
     */
    domain?: string;
    /**
     * 
     * @type {AdminDto}
     * @memberof RegisterCompanyRequestDto
     */
    admin?: AdminDto;
}

/**
 * Check if a given object implements the RegisterCompanyRequestDto interface.
 */
export function instanceOfRegisterCompanyRequestDto(value: object): value is RegisterCompanyRequestDto {
    return true;
}

export function RegisterCompanyRequestDtoFromJSON(json: any): RegisterCompanyRequestDto {
    return RegisterCompanyRequestDtoFromJSONTyped(json, false);
}

export function RegisterCompanyRequestDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): RegisterCompanyRequestDto {
    if (json == null) {
        return json;
    }
    return {
        
        'companyName': json['companyName'] == null ? undefined : json['companyName'],
        'domain': json['domain'] == null ? undefined : json['domain'],
        'admin': json['admin'] == null ? undefined : AdminDtoFromJSON(json['admin']),
    };
}

export function RegisterCompanyRequestDtoToJSON(json: any): RegisterCompanyRequestDto {
    return RegisterCompanyRequestDtoToJSONTyped(json, false);
}

export function RegisterCompanyRequestDtoToJSONTyped(value?: RegisterCompanyRequestDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'companyName': value['companyName'],
        'domain': value['domain'],
        'admin': AdminDtoToJSON(value['admin']),
    };
}